refactor: integrate profiling chunk logic into existing Scheduler
Remove separate ProfilingChunkScheduler subclass and add the dynamic
chunk sizing logic directly into the base Scheduler class. The subclass
was only needed in vllm-ascend as a patch mechanism; since we can modify
the Scheduler directly, a separate class is unnecessary.

Changes:
- scheduler.py: add profiling_chunk_manager init, time_budget in
  schedule(), run_profiling_chunk_init and helper methods
- Delete scheduler_profiling_chunk.py
- Update tests to use Scheduler directly
